backend/pyproject.toml
[tool.poetry]
name = "ai-tts-multilingual"
version = "0.1.0"
description = "AI-powered multilingual text-to-speech platform"
authors = ["Param Patel"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
TTS = "^0.22.0"  # Coqui TTS
torch = "^2.1.0"
torchaudio = "^2.1.0"
transformers = "^4.36.0"
phonemizer = "^3.2.1"
librosa = "^0.10.1"
soundfile = "^0.12.1"
numpy = "^1.24.0"
pyyaml = "^6.0.1"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.12.0"
isort = "^5.13.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
